moved ip parser and name generator to utils folder
